minor bug divide doubles not int's

+++ src/gr/auth/ee/lcs/classifiers/statistics/MeanAttributeSpecificityStatistic.java
@@ -42,8 +42,8 @@ public class MeanAttributeSpecificityStatistic implements ILCSMetric {
 			}
 		}
 
-		return specificAttibutes
-				/ (numberOfMacroclassifiers * numberOfAttributes);
+		return ((double)specificAttibutes)
+				/ ((double) (numberOfMacroclassifiers * numberOfAttributes));
 	}
 
 	/*
